+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+# 1.0.0 (2020-03-10)
+
+
+### Bug Fixes
+
+* **changelog:** 优化changelog ([d531833](https://github.com/Jennifer-Liuu/algorithms/commit/d5318332d237b38a84a333d4cae78acff08d2234))
+
+
+### Features
+
+* **array:** three sum ([ba8ccd4](https://github.com/Jennifer-Liuu/algorithms/commit/ba8ccd415e6e80d1adac5e1e83e7af8cdedf6faa))
+* **array & recall:** 字谜分组&回溯算法 ([7c6c484](https://github.com/Jennifer-Liuu/algorithms/commit/7c6c484d7e5a4ce57a5573777132f51f281d75e9))
+* **changelog:** format changelog ([3fc8dab](https://github.com/Jennifer-Liuu/algorithms/commit/3fc8dab84b4a15a02d95508ed3701d2e9436c0d5))
+* **linklist:** addTwoNumbers ([5121010](https://github.com/Jennifer-Liuu/algorithms/commit/5121010b1c01596e551006a84926a0b7a8ca8ee7))
+* **tree:** 二叉树中序遍历 ([09ee3ed](https://github.com/Jennifer-Liuu/algorithms/commit/09ee3edcc327e32ae10ea2571414c52a6e81582e))
+
+
+
 # 1.0.0 (2020-03-07)